requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
pydantic==1.10.0
pymongo==3.12.0
pymongo[srv]
twilio>=6.0.0,<7.0.0
tensorflow>=2.7.0,<2.16.0
keras>=2.7.0
openai
python-jose
passlib
python-decouple
python-dotenv
dnspython
pymongo





